4.21*: bloglist expansion, step9

+++ part4/bloglist/controllers/blogs.js
@@ -36,6 +36,13 @@ blogsRouter.post('/', async (req, res) => {
 })
 
 blogsRouter.delete('/:id', async (req, res) => {
+    const decodedToken = jwt.verify(req.token, process.env.SECRET)
+    const blog = await Blog.findById(req.params.id)
+
+    if(!req.token || blog.user.toString() !== decodedToken.id.toString()){
+        return res.status(401).json({ error: 'token missing or invalid user' })
+    }
+
     await Blog.findByIdAndRemove(req.params.id)
     res.status(204).end()
 })
